Implement audio buffering based on desired latency (50 ms currently)
This mostly gets rid of stuttering that occurs due to slight timing differences between the rate at which the APU generates samples and the audio playback rate.

+++ src/Apu.h
@@ -37,6 +37,6 @@ private:
 	std::shared_ptr<PulseChannel> m_pulseChannels[2];
 	std::shared_ptr<TriangleChannel> m_triangleChannel;
 	std::shared_ptr<NoiseChannel> m_noiseChannel;
-	float32 m_elapsedCpuCycles;
+	float64 m_elapsedCpuCycles;
 	std::shared_ptr<AudioDriver> m_audioDriver;
 };
